% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACB.R
\name{ACB_suploss}
\alias{ACB_suploss}
\alias{ACB}
\title{Calculate capital gains from realized gain transactions}
\usage{
ACB_suploss(
  data,
  transaction = "transaction",
  price = "price",
  quantity = "quantity",
  fees = "fees",
  total.price = "total.price",
  spot.rate = "spot.rate",
  as.revenue = c("staking", "interest", "mining")
)

ACB(
  data,
  transaction = "transaction",
  price = "price",
  quantity = "quantity",
  fees = "fees",
  total.price = "total.price",
  spot.rate = "spot.rate",
  as.revenue = c("staking", "interest", "mining")
)
}
\arguments{
\item{data}{The dataframe}

\item{transaction}{Name of transaction column}

\item{price}{Name of price column}

\item{quantity}{Name of quantity column}

\item{fees}{Name of fees column}

\item{total.price}{Name of total.price column}

\item{spot.rate}{Name of spot.rate column}

\item{as.revenue}{Name of as.revenue column}
}
\description{
Calculate realized and unrealized capital gains/losses
}
\examples{
\dontrun{
ACB(data)
}
}
\keyword{crypto}
\keyword{money}
