% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustedcostbase.ca.R
\docType{data}
\name{adjustedcostbase.ca}
\alias{adjustedcostbase.ca}
\title{Sample data set provided by adjustedcostbase.ca}
\format{
A data frame with 4 rows and 4 variables:
\describe{
\item{transaction}{buy or sell}
\item{price}{in dollars}
\item{quantity}{how much of the stock/coin}
\item{fees}{any transaction fees}
...
}
}
\source{
\url{https://www.adjustedcostbase.ca/blog/how-to-calculate-adjusted-cost-base-acb-and-capital-gains/}
}
\usage{
adjustedcostbase.ca
}
\description{
A dataset from adjustedcostbase.ca to demonstrate adjusted cost base as well as
capital gains/losses. Used as demo for our own gains() function.
}
\keyword{datasets}
