% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustedcostbase.ca.R
\docType{data}
\name{adjustedcostbase.ca1}
\alias{adjustedcostbase.ca1}
\alias{adjustedcostbase.ca2}
\alias{adjustedcostbase.ca3}
\alias{adjustedcostbase.ca4}
\alias{adjustedcostbase.ca5}
\alias{adjustedcostbase.ca6}
\title{Sample data sets provided by adjustedcostbase.ca}
\format{
Data frames with 4 variables:
\describe{
\item{date}{the date}
\item{transaction}{buy or sell}
\item{quantity}{how much of the stock/coin}
\item{price}{the spot rate, in dollars}
\item{fees}{any transaction fees}
...
}

An object of class \code{data.frame} with 4 rows and 4 columns.

An object of class \code{data.frame} with 4 rows and 4 columns.

An object of class \code{data.frame} with 3 rows and 4 columns.

An object of class \code{data.frame} with 2 rows and 4 columns.

An object of class \code{data.frame} with 5 rows and 4 columns.
}
\source{
\url{https://www.adjustedcostbase.ca/blog/how-to-calculate-adjusted-cost-base-acb-and-capital-gains/, https://www.adjustedcostbase.ca/blog/what-is-the-superficial-loss-rule/}
}
\usage{
adjustedcostbase.ca1

adjustedcostbase.ca2

adjustedcostbase.ca3

adjustedcostbase.ca4

adjustedcostbase.ca5

adjustedcostbase.ca6
}
\description{
Data sets from adjustedcostbase.ca to demonstrate adjusted cost base as well as
capital gains/losses. Used as demo for our own \code{\link[=ACB]{ACB()}} function.
}
\keyword{datasets}
