% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustedcostbase.ca.R
\docType{data}
\name{adjustedcostbase.ca_1}
\alias{adjustedcostbase.ca_1}
\alias{adjustedcostbase.ca_2}
\alias{adjustedcostbase.ca_3}
\alias{adjustedcostbase.ca_4}
\alias{adjustedcostbase.ca_5}
\alias{adjustedcostbase.ca_6}
\title{Sample data sets provided by adjustedcostbase.ca}
\format{
A data frame with 4 rows and 4 variables:
\describe{
\item{date}{the date}
\item{transaction}{buy or sell}
\item{quantity}{how much of the stock/coin}
\item{price}{the spot rate, in dollars}
\item{fees}{any transaction fees}
...
}

An object of class \code{data.frame} with 4 rows and 4 columns.

An object of class \code{data.frame} with 4 rows and 4 columns.

An object of class \code{data.frame} with 3 rows and 4 columns.

An object of class \code{data.frame} with 2 rows and 4 columns.

An object of class \code{data.frame} with 5 rows and 4 columns.
}
\source{
\url{https://www.adjustedcostbase.ca/blog/how-to-calculate-adjusted-cost-base-acb-and-capital-gains/}
}
\usage{
adjustedcostbase.ca_1

adjustedcostbase.ca_2

adjustedcostbase.ca_3

adjustedcostbase.ca_4

adjustedcostbase.ca_5

adjustedcostbase.ca_6
}
\description{
A dataset from adjustedcostbase.ca to demonstrate adjusted cost base as well as
capital gains/losses. Used as demo for our own gains() function.
}
\keyword{datasets}
