% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_profit.R
\name{calculate_profit_fifo}
\alias{calculate_profit_fifo}
\title{Calculate Profit from Trades using FIFO Principle}
\usage{
calculate_profit_fifo(trades)
}
\arguments{
\item{trades}{A dataframe of trades, which should have at least the following columns:
'time' (the time of the trade), 'pair' (the trading pair), 'type' (the type of trade - buy or sell),
'vol' (the volume of the trade), and 'price' (the price per unit).}
}
\value{
A dataframe summarizing the profit. Each row represents a profit calculated from a
"sell" trade, with columns for 'time' (the time of the trade), 'pair' (the trading pair),
and 'profit' (the profit from the trade).
}
\description{
This function calculates the profit for each trade using the First-In-First-Out (FIFO) principle,
given a dataframe of trades. For each unique pair of assets, the function calculates the profit
for each "sell" trade based on the price difference with the earliest (first-in) "buy" trades.
}
