% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing_transactions.R
\name{check_missing_transactions}
\alias{check_missing_transactions}
\title{View rows with negative total balances}
\usage{
check_missing_transactions(formatted.ACB)
}
\arguments{
\item{formatted.ACB}{The formatted.ACB file}
}
\value{
A data frame, filtered with the rows containing a total
quantity smaller than zero.
}
\description{
View rows with negative total balances to help identify
missing transactions.
}
\examples{
all.data <- format_shakepay(data_shakepay)
formatted.ACB <- format_ACB(all.data, verbose = FALSE)
check_missing_transactions(formatted.ACB)
}
