% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypto2fiat.R
\name{crypto2fiat}
\alias{crypto2fiat}
\title{Convert a cryptocurrency value to fiat (e.g., CAD)}
\usage{
crypto2fiat(coin, date, currency = "CAD")
}
\arguments{
\item{coin}{The coin}

\item{date}{The date}

\item{currency}{The currency}
}
\description{
Format a .csv file from Newton for later ACB processing.
}
\examples{
\dontrun{
crypto2fiat(data)
}

}
\keyword{crypto}
\keyword{money}
