% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_adjustedcostbase.R
\docType{data}
\name{data_adjustedcostbase1}
\alias{data_adjustedcostbase1}
\alias{data_adjustedcostbase2}
\alias{data_adjustedcostbase3}
\alias{data_adjustedcostbase4}
\alias{data_adjustedcostbase5}
\alias{data_adjustedcostbase6}
\title{Sample data sets provided by adjustedcostbase.ca}
\format{
Data frames with 4 variables:
\describe{
\item{date}{the date}
\item{transaction}{buy or sell}
\item{quantity}{how much of the stock/coin}
\item{price}{the spot rate, in dollars}
\item{fees}{any transaction fees}
...
}

An object of class \code{data.frame} with 4 rows and 4 columns.

An object of class \code{data.frame} with 4 rows and 4 columns.

An object of class \code{data.frame} with 3 rows and 4 columns.

An object of class \code{data.frame} with 2 rows and 4 columns.

An object of class \code{data.frame} with 5 rows and 4 columns.
}
\usage{
data_adjustedcostbase1

data_adjustedcostbase2

data_adjustedcostbase3

data_adjustedcostbase4

data_adjustedcostbase5

data_adjustedcostbase6
}
\description{
Data sets from adjustedcostbase.ca to demonstrate adjusted cost base as well as
capital gains/losses. Used as demo for our own \code{\link[=ACB]{ACB()}} function.
}
\references{
\itemize{
\item https://www.adjustedcostbase.ca/blog/how-to-calculate-adjusted-cost-base-acb-and-capital-gains/
\item https://www.adjustedcostbase.ca/blog/what-is-the-superficial-loss-rule/
}
}
\keyword{datasets}
