% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_binance.R
\docType{data}
\name{data_binance}
\alias{data_binance}
\title{Sample data set of a fictive Binance transaction history file}
\format{
A data frame with 39 rows and 7 variables:
\describe{
\item{User_ID}{user ID}
\item{UTC_Time}{the date}
\item{Account}{account type}
\item{Operation}{transaction type}
\item{Coin}{currency}
\item{Change}{quantity}
\item{Remark}{comment}
...
}
}
\source{
\url{https://www.binance.com/}
}
\usage{
data_binance
}
\description{
A fictive Binance data set to demonstrate \code{\link[=data_binance]{data_binance()}}.
}
\keyword{datasets}
