% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_binance_withdrawals.R
\docType{data}
\name{data_binance_withdrawals}
\alias{data_binance_withdrawals}
\title{Sample data set of a fictive Binance withdrawals transaction history file}
\format{
A data frame with 3 rows and 10 variables:
\describe{
\item{Date(UTC)}{the date}
\item{Coin}{currency}
\item{Network}{network type}
\item{Amount}{quantity}
\item{TransactionFee}{transaction fee}
\item{Address}{destination address}
\item{TXID}{transaction ID}
\item{SourceAddress}{source address}
\item{PaymentID}{payment ID}
\item{Status}{status}
...
}
}
\source{
\url{https://www.binance.com/}
}
\usage{
data_binance_withdrawals
}
\description{
A fictive withdrawals Binance data set to demonstrate \code{\link[=format_binance_withdrawals]{format_binance_withdrawals()}}.
}
\keyword{datasets}
