% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_blockfi.R
\docType{data}
\name{data_blockfi}
\alias{data_blockfi}
\title{Sample data set of a fictive BlockFi transaction history file}
\format{
A data frame with 15 rows and 4 variables:
\describe{
\item{Cryptocurrency}{currency}
\item{Amount}{quantity}
\item{Transaction.Type}{transaction type}
\item{Confirmed.At}{the date}
...
}
}
\source{
\url{https://blockfi.com/}
}
\usage{
data_blockfi
}
\description{
A fictive BlockFi data set to demonstrate \code{\link[=format_blockfi]{format_blockfi()}}.
}
\keyword{datasets}
