% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_CDC_exchange_rewards.R
\docType{data}
\name{data_CDC_exchange_rewards}
\alias{data_CDC_exchange_rewards}
\title{Sample data set of a fictive Crypto.com exchange rewards transaction history file}
\format{
A data frame with 10 rows and 12 variables:
\describe{
\item{Date}{the date}
\item{Sent.Amount}{sent quantity}
\item{Sent.Currency}{sent currency}
\item{Received.Amount}{received quantity}
\item{Received.Currency}{received currency}
\item{Fee.Amount}{fee quantity}
\item{Fee.Currency}{fee currency}
\item{Net.Worth.Amount}{net worth quantity}
\item{Net.Worth.Currency}{net worth currency}
\item{Label}{transaction type}
\item{Description}{description}
\item{TxHash}{transaction hash}
...
}
}
\source{
\url{https://crypto.com/exchange/}
}
\usage{
data_CDC_exchange_rewards
}
\description{
A fictive Crypto.com exchange rewards data set to demonstrate \code{\link[=format_CDC_exchange_rewards]{format_CDC_exchange_rewards()}}.
}
\keyword{datasets}
