% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_CDC_exchange_trades.R
\docType{data}
\name{data_CDC_exchange_trades}
\alias{data_CDC_exchange_trades}
\title{Sample data set of a fictive Crypto.com exchange trades transaction history file}
\format{
A data frame with 7 rows and 10 variables:
\describe{
\item{Order.ID}{order id}
\item{Trade.ID}{trade id}
\item{Time..UTC.}{the date}
\item{Symbol}{trade pair}
\item{Side}{buyer or seller side}
\item{Trade.Price}{trade price}
\item{Trade.Amount}{trade quantity}
\item{Volume.of.Business}{volume of business}
\item{Fee}{fee}
\item{Fee.Currency}{fee currency}
...
}
}
\source{
\url{https://crypto.com/exchange/}
}
\usage{
data_CDC_exchange_trades
}
\description{
A fictive Crypto.com exchange trades data set to demonstrate \code{\link[=format_CDC_exchange_trades]{format_CDC_exchange_trades()}}.
}
\keyword{datasets}
