% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_celsius.R
\docType{data}
\name{data_celsius}
\alias{data_celsius}
\title{Sample data set of a fictive Celsius transaction history file}
\format{
A data frame with 12 rows and 9 variables:
\describe{
\item{Internal.id}{internal id}
\item{Date.and.time}{the date}
\item{Transaction.type}{transaction type}
\item{Coin.type}{coin type}
\item{Coin.amount}{quantity}
\item{USD.Value}{value in USD}
\item{Original.Reward.Coin}{original reward coin}
\item{Reward.Amount.In.Original.Coin}{reward amount in original coin}
\item{Confirmed}{status of transaction}
...
}
}
\source{
\url{https://celsius.network/}
}
\usage{
data_celsius
}
\description{
A fictive Celsius data set to demonstrate \code{\link[=format_celsius]{format_celsius()}}.
}
\keyword{datasets}
