% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_coinbase.R
\docType{data}
\name{data_coinbase}
\alias{data_coinbase}
\title{Sample data set of a fictive Coinbase transaction history file}
\format{
A data frame with 5 rows and 10 variables:
\describe{
\item{Timestamp}{the date}
\item{Transaction.Type}{transaction type}
\item{Asset}{coin}
\item{Quantity.Transacted}{quantity}
\item{Spot.Price.Currency}{currency of spot price}
\item{Spot.Price.at.Transaction}{spot price}
\item{Subtotal}{subtotal}
\item{Total..inclusive.of.fees.and.or.spread.}{Grand total}
\item{Fees.and.or.Spread}{fees}
\item{Notes}{comments}
...
}
}
\source{
\url{https://www.coinbase.com/}
}
\usage{
data_coinbase
}
\description{
A fictive Coinbase data set to demonstrate \code{\link[=format_coinbase]{format_coinbase()}}.
}
\keyword{datasets}
