% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_coinpanda.R
\docType{data}
\name{data_coinpanda1}
\alias{data_coinpanda1}
\alias{data_coinpanda2}
\title{Sample data sets provided by coinpanda.io}
\format{
A data frame with 4 rows and 6 variables:
\describe{
\item{type}{type of transaction}
\item{date}{the date}
\item{currency}{the coin}
\item{amount}{quantity}
\item{price}{the total.price}
\item{fees}{any transaction fees}
...
}

An object of class \code{data.frame} with 4 rows and 6 columns.
}
\source{
\url{https://coinpanda.io/blog/crypto-taxes-canada-adjusted-cost-base/}
}
\usage{
data_coinpanda1

data_coinpanda2
}
\description{
Data sets from coinpanda.io to demonstrate adjusted cost base as well as
capital gains/losses. Used as demo for our own \code{\link[=ACB]{ACB()}} function.
}
\keyword{datasets}
