% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_coinsmart.R
\docType{data}
\name{data_coinsmart}
\alias{data_coinsmart}
\title{Sample data set of a fictive CoinSmart transaction history file}
\format{
A data frame with 8 rows and 7 variables:
\describe{
\item{Credit}{amount added}
\item{Debit}{amount removed}
\item{TransactionType}{transaction type}
\item{ReferenceType}{description}
\item{Product}{the currency}
\item{Balance}{the current balance}
\item{TimeStamp}{the date}
...
}
}
\source{
\url{https://www.coinsmart.com/}
}
\usage{
data_coinsmart
}
\description{
A fictive CoinSmart data set to demonstrate \code{\link[=format_coinsmart]{format_coinsmart()}}.
}
\keyword{datasets}
