% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_exodus.R
\docType{data}
\name{data_exodus}
\alias{data_exodus}
\title{Sample data set of a fictive Exodus wallet transaction history file}
\format{
A data frame with 8 rows and 17 variables:
\describe{
\item{DATE}{the date}
\item{TYPE}{transaction type}
\item{FROMPORTFOLIO}{from where}
\item{TOPORTFOLIO}{to where}
\item{OUTAMOUNT}{quantity sent}
\item{OUTCURRENCY}{currency sent}
\item{FEEAMOUNT}{the fees}
\item{FEECURRENCY}{fee currency}
\item{TOADDRESS}{address sent to}
\item{OUTTXID}{transaction ID for out transactions}
\item{OUTTXURL}{transaction URL for out transactions}
\item{INAMOUNT}{quantity received}
\item{INCURRENCY}{currency received}
\item{INTXID}{transaction ID for in transactions}
\item{INTXURL}{transaction URL for in transactions}
\item{ORDERID}{order ID}
\item{PERSONALNOTE}{personal notes}
...
}
}
\source{
\url{https://www.exodus.com/}
}
\usage{
data_exodus
}
\description{
A fictive Exodus wallet data set to demonstrate \code{\link[=format_exodus]{format_exodus()}}.
}
\keyword{datasets}
