% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generic.R
\docType{data}
\name{data_generic1}
\alias{data_generic1}
\alias{data_generic2}
\alias{data_generic3}
\alias{data_generic4}
\title{Sample data sets of fictive generic transaction history files}
\format{
A data frame with 3 rows and 6-8 variables:
\describe{
\item{Date}{the date}
\item{Currency}{the currency}
\item{Quantity}{quantity}
\item{Spot.Rate}{the spot rate, in dollars}
\item{Total.Price}{the cost of the transaction}
\item{Transaction}{the type of transaction}
\item{Fees}{the transaction fees}
\item{Exchange}{the exchange}
...
}

An object of class \code{data.frame} with 3 rows and 7 columns.

An object of class \code{data.frame} with 3 rows and 7 columns.

An object of class \code{data.frame} with 3 rows and 6 columns.
}
\usage{
data_generic1

data_generic2

data_generic3

data_generic4
}
\description{
Fictive generic data sets to demonstrate \code{\link[=format_generic]{format_generic()}}.
}
\keyword{datasets}
