% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_koinly.R
\docType{data}
\name{data_koinly}
\alias{data_koinly}
\title{Sample data sets provided by koinly.io}
\format{
A data frame with 4 rows and 4 variables:
\describe{
\item{date}{the date}
\item{transaction}{buy or sell}
\item{currency}{the coin}
\item{quantity}{how much of the stock/coin}
\item{spot.rate}{the price of the coin}
...
}
}
\source{
\url{https://koinly.io/blog/calculating-crypto-taxes-canada/}
}
\usage{
data_koinly
}
\description{
A data set from koinly.io to demonstrate adjusted cost base as well as
capital gains/losses. Used as demo for our own \code{\link[=ACB]{ACB()}} function.
}
\keyword{datasets}
