% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_newton.R
\docType{data}
\name{data_newton}
\alias{data_newton}
\title{Sample data set of a fictive Newton transaction history file}
\format{
A data frame with 12 rows and 3 variables:
\describe{
\item{Date}{the date}
\item{Type}{transaction type}
\item{Received.Quantity}{received quantity}
\item{Received.Currency}{received currency}
\item{Sent.Quantity}{sent quantity}
\item{Sent.Currency}{sent currency}
\item{Fee.Amount}{fee amount}
\item{Fee.Currency}{fee currency}
\item{Tag}{the tag}
...
}
}
\source{
\url{https://www.newton.co/}
}
\usage{
data_newton
}
\description{
A fictive Newton data set to demonstrate \code{\link[=format_newton]{format_newton()}}.
}
\keyword{datasets}
