% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_presearch.R
\docType{data}
\name{data_presearch}
\alias{data_presearch}
\title{Sample data set of a fictive Presearch transaction history file}
\format{
A data frame with 12 rows and 3 variables:
\describe{
\item{date}{the date}
\item{amount}{quantity}
\item{description}{description}
...
}
}
\source{
\url{https://presearch.com/}
}
\usage{
data_presearch
}
\description{
A fictive Presearch data set to demonstrate \code{\link[=format_presearch]{format_presearch()}}.
}
\keyword{datasets}
