% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_shakepay.R
\docType{data}
\name{data_shakepay}
\alias{data_shakepay}
\title{Sample data set of a fictive Shakepay transaction history file}
\format{
A data frame with 8 rows and 11 variables:
\describe{
\item{Date}{the date}
\item{Amount.Debited}{amount debited}
\item{Asset.Debited}{currency debited}
\item{Amount.Credited}{amount credited}
\item{Asset.Credited}{currency credited}
\item{Market.Value}{market value}
\item{Market.Value.Currency}{market value currency}
\item{Book.Cost}{value of transaction (CAD)}
\item{Book.Cost.Currency}{currency of Book.Cost (CAD)}
\item{Type}{transaction type}
\item{Spot.Rate}{transaction spot rate on Shakepay}
\item{Buy...Sell.Rate}{Buy / sell rate}
\item{Description}{Buy, Sell, Send, or Reward}
}
}
\source{
\url{https://shakepay.com/}
}
\usage{
data_shakepay
}
\description{
A fictive Shakepay data set to demonstrate \code{\link[=format_shakepay]{format_shakepay()}}.
}
\keyword{datasets}
