% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_export_csv}
\alias{extract_export_csv}
\title{Unzip and extract CSV file}
\usage{
extract_export_csv(bin_output)
}
\arguments{
\item{bin_output}{Binary output data}
}
\value{
The extracted CSV data as a data frame
}
\description{
Unzip and extract CSV file
}
