% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_cronos_pos.R
\name{fetch_cronos_pos}
\alias{fetch_cronos_pos}
\title{Download transaction data from the Cronos POS chain}
\usage{
fetch_cronos_pos(address, limit = 100, perform = TRUE, verbose = TRUE)
}
\arguments{
\item{address}{The Cronos POS wallet address (starts with "cro1...")}

\item{limit}{Query number of transactions results per page returned
(default 100)}

\item{perform}{Whether to execute the API request (if \code{TRUE}), or just return
how the request would be formatted (if \code{FALSE}).}

\item{verbose}{Logical, if \code{FALSE}, does not print warnings to console.}
}
\value{
A data frame of exchange transactions, formatted for further processing.
}
\description{
Downloads transaction data from the Cronos Proof-of-Stake
chain / the Crypto.com DeFi wallet, through the API.
}
\examples{
fetch_cronos_pos(limit = 10, perform = FALSE,
 "cro1dec64zlzracgz7fs4thzx45q7a48s22d4ll8m6")
}
