% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_prices.R
\name{fetch_prices}
\alias{fetch_prices}
\title{Fetch prices of daily transactions}
\usage{
fetch_prices(coin, data, coins.list)
}
\arguments{
\item{coin}{The coin}

\item{data}{The dataframe}

\item{coins.list}{The coins list}
}
\description{
Fetches daily prices of coins through the crypto2 package.
}
\examples{
\dontrun{
fetch_prices(data)
}
}
\keyword{crypto}
\keyword{money}
