% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_binance_earn.R
\name{format_binance_earn}
\alias{format_binance_earn}
\title{Format Binance earn file}
\usage{
format_binance_earn(data, filetype = "all")
}
\arguments{
\item{data}{The dataframe}

\item{filetype}{Which file type to use, one of ("all", for all
transactions, or "earn", for just earn).}
}
\description{
Format a .csv earn history file from Binance for later
ACB processing.
}
\details{
There are two ways to get this file. The first,
recommended option, is to download your overall transaction report
(this will also include e.g., "Referral Kickback" rewards).
To get this file, connect to your Binance account on desktop, click
"Wallet" (top right), "Transaction History", then in the top-right,
"Generate all statements". For "Time", choose "Customized" and pick
your time frame.

For option two, click "Orders" (top right), "Earn History", then click on
"Export". In "Type", choose "Interest", and choose your time frame
(you will probably need to choose "Beyond 6 months - Custom").
In both cases, you are allowed one year of transactions. If you have more,
you might have to download more than one file and merge them before
using this function.

Warning: This does NOT process TRADES or WITHDRAWALS (see the
\code{format_binance_trades()} and \code{format_binance_withdrawals()}
functions for this purpose).
}
\examples{
\dontrun{
format_binance_earn(data)
}
}
\keyword{crypto}
\keyword{money}
