% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_binance_withdrawals.R
\name{format_binance_withdrawals}
\alias{format_binance_withdrawals}
\title{Format Binance withdrawal file}
\usage{
format_binance_withdrawals(data)
}
\arguments{
\item{data}{The dataframe}
}
\description{
Format a .xlsx withdrawal history file from Binance for later ACB processing.

Warning: This does NOT process TRADES See the format_binance_trades() function for this purpose.
}
\examples{
\dontrun{
format_binance_withdrawals(data)
}
}
\keyword{crypto}
\keyword{money}
