% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_blockfi.R
\name{format_blockfi}
\alias{format_blockfi}
\title{Format BlockFi file}
\usage{
format_blockfi(data)
}
\arguments{
\item{data}{The dataframe}
}
\description{
Format a .csv transaction history file from BlockFi for later ACB processing.
}
\examples{
\dontrun{
format_blockfi(data)
}
}
\keyword{crypto}
\keyword{money}
