% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_BSC.R
\name{format_BSC}
\alias{format_BSC}
\title{Format custom Binance Smart Chain file}
\usage{
format_BSC(data)
}
\arguments{
\item{data}{The dataframe}
}
\description{
Format a custom .csv transaction history file from the Binance Smart Chain wallet for later ACB processing.
}
\examples{
\dontrun{
format_BSC(data)
}
}
\keyword{crypto}
\keyword{money}
