% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_CDC.R
\name{format_CDC}
\alias{format_CDC}
\title{Format Crypto.com App file}
\usage{
format_CDC(data)
}
\arguments{
\item{data}{The dataframe}
}
\description{
Format a .csv transaction history file from Crypto.com for later ACB processing.
}
\examples{
\dontrun{
format_CDC(data)
}
}
\keyword{crypto}
\keyword{money}
