% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_CDC_exchange_trades.R
\name{format_CDC_exchange_trades}
\alias{format_CDC_exchange_trades}
\title{Format CDC exchange file (FOR TRADES ONLY)}
\usage{
format_CDC_exchange_trades(data)
}
\arguments{
\item{data}{The dataframe}
}
\description{
Format a .csv transaction history file from the Crypto.com exchange for later ACB processing. Only processes trades, not rewards (see \code{format_CDC_exchange_rewards} for this).

Original file name of the right file from the exchange is called "SPOT_TRADE.csv", make sure you have the right one.
}
\examples{
\dontrun{
format_CDC_exchange_trades(data)
}
}
\keyword{crypto}
\keyword{money}
