% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_coinbase.R
\name{format_coinbase}
\alias{format_coinbase}
\title{Format Coinbase file}
\usage{
format_coinbase(data)
}
\arguments{
\item{data}{The dataframe}
}
\value{
A data frame of exchange transactions, formatted for further processing.
}
\description{
Format a .csv transaction history file from Coinbase for later ACB
processing.
}
\examples{
format_coinbase(data_coinbase)
}
