% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_coinsmart.R
\name{format_coinsmart}
\alias{format_coinsmart}
\title{Format CoinSmart file}
\usage{
format_coinsmart(data, list.prices = NULL, force = FALSE)
}
\arguments{
\item{data}{The dataframe}

\item{list.prices}{An optional explicit \code{list.prices} object from which to
fetch coin prices. For exchanges that require external pricing, it must
contain at least \code{currency}, \code{spot.rate2}, and \code{date2}.}

\item{force}{Whether to force recreating \code{list.prices} even though
it already exists (e.g., if you added new coins or new dates).}
}
\value{
A data frame of exchange transactions, formatted for further processing.
}
\description{
Format a .csv transaction history file from CoinSmart for later ACB processing.
}
\examples{
format_coinsmart(data_coinsmart)
}
