% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_cronos_pos.R
\name{format_cronos_pos}
\alias{format_cronos_pos}
\title{Format transaction data from the Cronos POS chain}
\usage{
format_cronos_pos(data, list.prices = NULL, force = FALSE)
}
\arguments{
\item{data}{The dataframe}

\item{list.prices}{A \code{list.prices} object from which to fetch coin prices.}

\item{force}{Whether to force recreating \code{list.prices} even though
it already exists (e.g., if you added new coins or new dates).}
}
\value{
A data frame of exchange transactions, formatted for further processing.
}
\description{
Format a .csv transaction history file from the Crypto.com DeFi
wallet for later ACB processing.

Use \code{\link[=fetch_cronos_pos]{fetch_cronos_pos()}} to first download the data.
}
\examples{
\donttest{
data <- fetch_cronos_pos(limit = 30,
 "cro1juv4wda4ty2tas8dwh7jc2ea73ewhtc26eyxwt")
format_cronos_pos(data)
}
}
