% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_dollars.R
\name{format_dollars}
\alias{format_dollars}
\title{Format numeric values to dollars}
\usage{
format_dollars(x, to = "character")
}
\arguments{
\item{x}{The formatted.ACB file}

\item{to}{What to convert to, with otions \code{c("character", "numeric")}.}
}
\value{
A value representing dollars, either as a formatted
character string or as a numeric value.
}
\description{
Format numeric values with comma for thousands separator.
Can be converted back from this format to numeric using \code{to = "numeric"}.
}
\examples{
x <- format_dollars(1010.92)
x
format_dollars(x, to = "numeric")
}
