% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_gemini.R
\name{format_gemini}
\alias{format_gemini}
\title{Format Gemini file}
\usage{
format_gemini(data, list.prices = NULL, force = FALSE)
}
\arguments{
\item{data}{The dataframe}

\item{list.prices}{An optional explicit \code{list.prices} object from which to
fetch coin prices. For exchanges that require external pricing, it must
contain at least \code{currency}, \code{spot.rate2}, and \code{date2}.}

\item{force}{Whether to force recreating \code{list.prices} even though
it already exists (e.g., if you added new coins or new dates).}
}
\value{
A data frame of exchange transactions, formatted for further processing.
}
\description{
Format a .csv transaction history file from Gemini for later ACB processing.
Open the xlsx data file using \code{readxl::read_excel()}.
}
\examples{
format_gemini(data_gemini)
}
