% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_newton.R
\name{format_newton}
\alias{format_newton}
\title{Format Newton file}
\usage{
format_newton(data)
}
\arguments{
\item{data}{The dataframe}
}
\description{
Format a .csv transaction history file from Newton for later ACB processing.
}
\examples{
\dontrun{
format_newton(data)
}
}
\keyword{crypto}
\keyword{money}
