% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_pooltool.R
\name{format_pooltool}
\alias{format_pooltool}
\title{Format ADA rewards from blockchain CSV}
\usage{
format_pooltool(data)
}
\arguments{
\item{data}{The dataframe}
}
\value{
A data frame of exchange transactions, formatted for further processing.
}
\description{
Format a .csv transaction history file from the Cardano PoolTool for later ACB processing. Instructions: Use https://pooltool.io/ click on "rewards data for taxes", search your ADA address, scroll to the bottom of the page, and use the export tool to export all transactions. Make sure to use the "Generic(CSV)" format.
}
\details{
This is necessary e.g., if you used the Exodus wallet which does not report
ADA rewards in its transaction history file.
}
\examples{
format_pooltool(data_pooltool)
}
