% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_presearch.R
\name{format_presearch}
\alias{format_presearch}
\title{Format Presearch wallet file}
\usage{
format_presearch(data)
}
\arguments{
\item{data}{The dataframe}
}
\description{
Format a .csv transaction history file from Presearch for later ACB processing.
}
\examples{
\dontrun{
format_presearch(data)
}
}
\keyword{crypto}
\keyword{money}
