% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_uphold.R
\name{format_uphold}
\alias{format_uphold}
\title{Format Uphold file}
\usage{
format_uphold(data, list.prices = NULL, force = FALSE)
}
\arguments{
\item{data}{The dataframe}

\item{list.prices}{A \code{list.prices} object from which to fetch coin prices.}

\item{force}{Whether to force recreating \code{list.prices} even though
it already exists (e.g., if you added new coins or new dates).}
}
\value{
A data frame of exchange transactions, formatted for further processing.
}
\description{
Format a .csv transaction history file from Uphold for later ACB processing.
}
\examples{
format_uphold(data_uphold)
}
