% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_wealthsimple.R
\name{format_wealthsimple}
\alias{format_wealthsimple}
\title{Format Wealthsimple file}
\usage{
format_wealthsimple(data)
}
\arguments{
\item{data}{The dataframe}
}
\description{
Format a .csv transaction history file from Wealthsimple for later ACB processing.
}
\examples{
\dontrun{
format_wealthsimple(data)
}
}
