% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kraken_api.R
\name{generate_export_report}
\alias{generate_export_report}
\title{Generate trades or ledger export report}
\usage{
generate_export_report(start_time, end_time, report_type, description)
}
\arguments{
\item{start_time}{Start time of the report}

\item{end_time}{End time of the report}

\item{report_type}{Type of the report, trades or ledgers}

\item{description}{Description of the report}
}
\value{
The generated report
}
\description{
Generate trades or ledger export report
}
