% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_kraken_signature}
\alias{get_kraken_signature}
\title{Get API Signature}
\usage{
get_kraken_signature(private_key, nonce, post_string, method_path)
}
\arguments{
\item{private_key}{The private key}

\item{nonce}{The nonce}

\item{post_string}{The post string}

\item{method_path}{The method path}
}
\value{
The API signature
}
\description{
Get API Signature
}
