% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latest_transactions.R
\name{get_latest_transactions}
\alias{get_latest_transactions}
\title{Get the date from the latest transactions per exchange}
\usage{
get_latest_transactions(formatted.ACB)
}
\arguments{
\item{formatted.ACB}{The formatted.ACB file}
}
\value{
A data frame with two columns: the name of the exchange as well
as the date of the latest transaction.
}
\description{
Get the date from the latest transactions per exchange
}
\examples{
all.data <- format_shakepay(data_shakepay)
formatted.ACB <- format_ACB(all.data, verbose = FALSE)
get_latest_transactions(formatted.ACB)
}
