% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_proceeds.R
\name{get_proceeds}
\alias{get_proceeds}
\title{Get proceeds of all sold coins}
\usage{
get_proceeds(formatted.ACB, tax.year, local.timezone = Sys.timezone())
}
\arguments{
\item{formatted.ACB}{The \code{formatted.ACB} object,}

\item{tax.year}{which year}

\item{local.timezone}{which time zone}
}
\value{
A data frame, with the following columns: type, proceeds, ACB.total,
gains.
}
\description{
Get proceeds of sold coins, ACB of sold coins, and resulting
difference between the two, capital gains.
}
\examples{
all.data <- format_shakepay(data_shakepay)
formatted.ACB <- format_ACB(all.data, verbose = FALSE)
get_proceeds(formatted.ACB, 2021)
}
