% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sup_losses.R
\name{get_sup_losses}
\alias{get_sup_losses}
\title{Get superficial loss amounts}
\usage{
get_sup_losses(formatted.ACB, tax.year, local.timezone = Sys.timezone())
}
\arguments{
\item{formatted.ACB}{The dataframe \code{formatted.ACB},}

\item{tax.year}{which year}

\item{local.timezone}{which time zone}
}
\value{
A data frame, with the following columns: currency, sup.loss.
}
\description{
Get superficial loss amounts
}
\examples{
all.data <- format_shakepay(data_shakepay)
formatted.ACB <- format_ACB(all.data, verbose = FALSE)
get_sup_losses(formatted.ACB, 2021)
}
