% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listby_coin.R
\name{listby_coin}
\alias{listby_coin}
\title{List transactions by coin}
\usage{
listby_coin(formatted.ACB)
}
\arguments{
\item{formatted.ACB}{The dataframe}
}
\value{
A list of formatted data frames, by coin.
}
\description{
Provides a list of transactions, separated by coin..
}
\examples{
all.data <- format_shakepay(data_shakepay)
formatted.ACB <- format_ACB(all.data, verbose = FALSE)
listby_coin(formatted.ACB)
}
