% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_prices.R
\name{match_prices}
\alias{match_prices}
\title{Get Fair Market Value (FMV) of transactions}
\usage{
match_prices(
  data,
  slug = NULL,
  start.date = "2021-01-01",
  list.prices = NULL,
  force = FALSE,
  verbose = TRUE,
  coins.list = NULL,
  coin_hist = NULL,
  USD2CAD.table = NULL
)
}
\arguments{
\item{data}{The dataframe}

\item{slug}{Your coins to match. You must use the long name, the "slug",
not the ticker, see \code{\link[=prepare_list_prices]{prepare_list_prices()}} for more details.}

\item{start.date}{What date to start reporting prices for.}

\item{list.prices}{A \code{list.prices} object from which to fetch coin prices.}

\item{force}{Whether to force recreating \code{list.prices} even though
it already exists (e.g., if you added new coins or new dates).}

\item{verbose}{Logical; whether to print progress messages.}

\item{coins.list}{Optional explicit output from \code{\link[crypto2:crypto_list]{crypto2::crypto_list()}}.}

\item{coin_hist}{Optional explicit historical price data to transform into
a \code{list.prices} object.}

\item{USD2CAD.table}{Optional explicit USD/CAD rate table to use when
converting USD-denominated history to CAD.}
}
\value{
A data frame, with the following added columns: spot.rate.
}
\description{
Matches prices obtained through the \code{prepare_list_prices()}
function with the transaction data frame.
}
\examples{
data <- format_shakepay(data_shakepay)[c(1:2)]
match_prices(data)
}
