% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_list_prices.R
\name{prepare_list_prices}
\alias{prepare_list_prices}
\title{Prepare the list of coins for prices}
\usage{
prepare_list_prices(coins, start.date, end.date = Sys.Date())
}
\arguments{
\item{coins}{Which coins to include in the list.}

\item{start.date}{What date to start reporting prices for.}

\item{end.date}{What date to end reporting prices for.}
}
\description{
Prepare the list of coins for prices.
}
\examples{
\dontrun{
match_prices(data)
}
}
\keyword{crypto}
\keyword{money}
