% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_report.R
\name{print_report}
\alias{print_report}
\title{Print full crypto tax report}
\usage{
print_report(
  formatted.ACB,
  list.prices,
  tax.year = "all",
  local.timezone = "America/Montreal",
  name
)
}
\arguments{
\item{formatted.ACB}{The \code{formatted.ACB} object.}

\item{list.prices}{A \code{list.prices} object from which to fetch coin prices.}

\item{tax.year}{The tax year desired.}

\item{local.timezone}{Which time zone to use for the date of the report.}

\item{name}{Name of the individual for the report.}
}
\value{
A list, containing the following objects: report.overview,
report.summary, proceeds, sup.losses, table.revenues, tax.box,
pie_exchange, pie_revenue.

An HTML page containing a crypto tax report.
}
\description{
Will output a full crypto tax report in HTML format, which can then be printed or saved as PDF.
}
\examples{
\donttest{
list.prices <- prepare_list_prices(coins = "BTC", start.date = "2021-01-01")
all.data <- format_shakepay(data_shakepay)
formatted.ACB <- format_ACB(all.data, verbose = FALSE)
print_report(formatted.ACB, list.prices = list.prices, 
             tax.year = 2021, name = "Mr. Cryptoltruist")
}
\dontshow{
unlink("full_report_2021.html")
}

}
