% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kraken_api.R
\name{retrieve_export_data}
\alias{retrieve_export_data}
\title{Retrieve export data}
\usage{
retrieve_export_data(report_id)
}
\arguments{
\item{report_id}{The ID of the report to retrieve}
}
\value{
Export data as dataframe
}
\description{
Retrieve export data
}
