% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sup_loss.R
\name{suploss_range}
\alias{suploss_range}
\title{Calculate superficial capital losses}
\usage{
suploss_range(date)
}
\arguments{
\item{date}{The date}
}
\description{
Calculate superficial capital losses to be substracted from total capital losses.
}
\examples{
\dontrun{
ACB(data)
}
}
\keyword{crypto}
\keyword{money}
