% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_table.R
\name{tax_table}
\alias{tax_table}
\title{Format tax table for final HTML report}
\usage{
tax_table(table, repeat.header = FALSE, type = 1)
}
\arguments{
\item{table}{The table to format}

\item{repeat.header}{Logical, whether to repeat headers at the bottom.}

\item{type}{Type of table, one of 1 (default), 2, or 3.}
}
\value{
A flextable object, with certain formatting features.
}
\description{
Format tax tables for the final rmd/html report.
}
\examples{
all.data <- format_shakepay(data_shakepay)
formatted.ACB <- format_ACB(all.data, verbose = FALSE)
x <- get_sup_losses(formatted.ACB, 2021)
tax_table(x)
}
