% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data.R
\name{transform_ledger_to_trades}
\alias{transform_ledger_to_trades}
\title{Transform Ledger to Trades Dataframe}
\usage{
transform_ledger_to_trades(ledger_df)
}
\arguments{
\item{ledger_df}{A dataframe representing the ledger data. This dataframe should have at least the
following columns: 'type', 'refid', 'time', 'asset', 'amount', and 'fee'.}
}
\value{
A dataframe summarizing the trades. Each row represents a trade, with columns for 'time'
(the time of the trade), 'pair', 'type' (the type of trade - buy or sell), 'cost'
(price * vol), 'vol' (the volume of the trade), 'price' (the EUR price per unit), and 'fee'
}
\description{
This function transforms a given ledger dataframe into a trades dataframe. It filters out non-trading
activities and identifies buy/sell trades to generate a summary of trades. The generated trades dataframe
is similar to the structure of a trades export from Kraken. The function is needed because kraken trades
export does not include trades made through the classic kraken app (non pro)
}
