% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USD2CAD.R
\name{USD2CAD_crypto2}
\alias{USD2CAD_crypto2}
\title{Convert USD to CAD (using \code{crypto2})}
\usage{
USD2CAD_crypto2(
  data,
  start.date = "2020-01-01",
  end.date = lubridate::now("UTC"),
  conversion = "USD",
  currency = "CAD",
  force = FALSE
)
}
\arguments{
\item{data}{The data}

\item{start.date}{What date to start reporting prices for.}

\item{end.date}{What date to end reporting prices for.}

\item{conversion}{What to convert to}

\item{currency}{What to convert from}

\item{force}{Whether to force recreating \code{list.prices} even though
it already exists (e.g., if you added new coins or new dates).}
}
\value{
A data frame, with the following columns: date, CAD.rate.
}
\description{
This function allows you to convert USD to CAD.
\code{crypto2}
\code{crypto2} now only return USD rates, so CAD rates are not available
anymore... See:
https://github.com/sstoeckl/crypto2/blob/master/NEWS.md#crypto-200

\emph{"\code{fiat_list()} has been modified and no longer delivers all available
currencies and precious metals (therefore only USD and Bitcoin are
available any more)."}
}
