% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USD2CAD.R
\name{USD2CAD_priceR}
\alias{USD2CAD_priceR}
\title{Convert USD to CAD (using \code{priceR})}
\usage{
USD2CAD_priceR(data, conversion = "USD", currency = "CAD")
}
\arguments{
\item{data}{The data}

\item{conversion}{What to convert to}

\item{currency}{What to convert from}
}
\value{
A data frame, with the following columns: date, CAD.rate.
}
\description{
This function allows you to convert USD to CAD.
\code{priceR} now requires an API key from https://exchangerate.host/...
See all details on the package author README here: https://github.com/stevecondylios/priceR
"Set up only takes a minute and is free for 100 requests per account per calendar month."
"Go to https://exchangerate.host/, create a free account, and replace
7e5e3140140bd8e4f4650cc41fc772c0 with your API key in the following, and run once per R session."
}
