% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypto_pie.R
\name{crypto_pie}
\alias{crypto_pie}
\title{Make a pie chart of your crypto revenues}
\usage{
crypto_pie(table.revenues, by = "exchange")
}
\arguments{
\item{table.revenues}{The revenue table to plot}

\item{by}{To plot by which element, one of \code{c("exchange", "revenue.type")}.}
}
\description{
Format a .csv file from Newton for later ACB processing.
}
\examples{
\dontrun{
crypto_pie(data)
}
}
\keyword{crypto}
\keyword{money}
