% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_adalite.R
\name{format_adalite}
\alias{format_adalite}
\title{Format Adalite wallet file}
\usage{
format_adalite(data)
}
\arguments{
\item{data}{The dataframe}
}
\description{
Format a .csv transaction history file from the Adalite wallet for later ACB processing.
}
\examples{
\dontrun{
format_adalite(data)
}
}
