% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_binance.R
\name{format_binance}
\alias{format_binance}
\title{Format Binance earn file}
\usage{
format_binance(data)
}
\arguments{
\item{data}{The dataframe}
}
\description{
Format a .csv earn history file from Binance for later
ACB processing.
}
\details{
To get this file. Download your overall transaction report
(this will include your trades, rewards, & "Referral Kickback" rewards).
To get this file, connect to your Binance account on desktop, click
"Wallet" (top right), "Transaction History", then in the top-right,
"Generate all statements". For "Time", choose "Customized" and pick
your time frame.

Warning: This does NOT process WITHDRAWALS (see the
\code{format_binance_withdrawals()} function for this purpose).
}
\examples{
\dontrun{
format_binance_earn(data)
}
}
