% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_binance_trades.R
\name{format_binance_trades}
\alias{format_binance_trades}
\title{Format Binance trades file}
\usage{
format_binance_trades(data)
}
\arguments{
\item{data}{The dataframe}
}
\description{
Format a .xlsx trades history file from Binance for later ACB processing.

Warning: This does NOT process WITHDRAWALS and associated fees. See the format_binance_withdrawals() function for this purpose.
}
\examples{
\dontrun{
format_binance_trades(data)
}
}
\keyword{crypto}
\keyword{money}
