% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_CDC.R
\name{format_CDC}
\alias{format_CDC}
\title{Format Crypto.com App file}
\usage{
format_CDC(data)
}
\arguments{
\item{data}{The dataframe}
}
\value{
A data frame of exchange transactions, formatted for further processing.
}
\description{
Format a .csv transaction history file from Crypto.com for later
ACB processing.
}
\details{
Be aware that CDC unfortunately does not include the withdrawal
fees in their exported transaction files (please lobby to include this feature).
This function attempts to guess some known withdrawal fees at some point in time
but depending on when the withdrawals were made, the withdrawal fees are most
certainly inaccurate. You will have to make a manual correction for the
withdrawal fees after using \code{format_CDC}, on the resulting dataframe.
}
\examples{
format_CDC(data_CDC)
}
