% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_celsius.R
\name{format_celsius}
\alias{format_celsius}
\title{Format Celsius file}
\usage{
format_celsius(data)
}
\arguments{
\item{data}{The dataframe}
}
\value{
A data frame of exchange transactions, formatted for further processing.
}
\description{
Format a .csv transaction history file from Celsius for later ACB processing.
}
\examples{
format_celsius(data_celsius)
}
