% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_coinsmart.R
\name{format_coinsmart}
\alias{format_coinsmart}
\title{Format CoinSmart file}
\usage{
format_coinsmart(data)
}
\arguments{
\item{data}{The dataframe}
}
\description{
Format a .csv transaction history file from CoinSmart for later ACB processing.
}
\examples{
\dontrun{
format_coinsmart(data)
}
}
