% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_exodus.R
\name{format_exodus}
\alias{format_exodus}
\title{Format Exodus wallet file}
\usage{
format_exodus(data)
}
\arguments{
\item{data}{The dataframe}
}
\description{
Format a .csv transaction history file from the Exodus wallet for later ACB processing.
}
\examples{
\dontrun{
format_exodus(data)
}
}
