% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_gemini.R
\name{format_gemini}
\alias{format_gemini}
\title{Format Gemini file}
\usage{
format_gemini(data)
}
\arguments{
\item{data}{The dataframe}
}
\description{
Format a .csv transaction history file from Gemini for later ACB processing.
}
\examples{
\dontrun{
format_gemini(data)
}
}
\keyword{crypto}
\keyword{money}
