% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_pooltool.R
\name{format_pooltool}
\alias{format_pooltool}
\title{Format ADA rewards from blockchain CSV}
\usage{
format_pooltool(data)
}
\arguments{
\item{data}{The dataframe}
}
\description{
Format a .csv transaction history file from the Cardano PoolTool for later ACB processing. Instructions: Use https://pooltool.io/ click on "rewards data for taxes", search your ADA address, scroll to the bottom of the page, and use the export tool to export all transactions. Make sure to use the "Generic(CSV)" format.
}
\examples{
\dontrun{
format_pooltool(data)
}
}
\keyword{crypto}
\keyword{money}
