% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_shakepay.R
\name{format_shakepay}
\alias{format_shakepay}
\title{Format Shakepay file}
\usage{
format_shakepay(data)
}
\arguments{
\item{data}{The dataframe}
}
\value{
A data frame of exchange transactions, formatted for further processing.
}
\description{
Format a .csv transaction history file from Shakepay for later ACB processing.
}
\examples{
formatted.shakepay <- format_shakepay(data_shakepay)
formatted.shakepay
}
