% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sup_loss.R
\name{format_suploss}
\alias{format_suploss}
\title{Calculate superficial capital losses}
\usage{
format_suploss(data, cl = NULL)
}
\arguments{
\item{data}{The data}

\item{cl}{The number of cores to use.}
}
\description{
Calculate superficial capital losses to be substracted from total capital losses.
}
\examples{
\dontrun{
ACB(data)
}
}
\keyword{crypto}
\keyword{money}
