% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_suploss.R
\name{format_suploss}
\alias{format_suploss}
\title{Calculate superficial capital losses}
\usage{
format_suploss(
  data,
  transaction = "transaction",
  quantity = "quantity",
  cl = NULL
)
}
\arguments{
\item{data}{The data}

\item{transaction}{Name of transaction column}

\item{quantity}{Name of quantity column}

\item{cl}{The number of cores to use.}
}
\description{
Calculate superficial capital losses to be substracted from total capital losses.
}
\examples{
data <- data_adjustedcostbase1
format_suploss(data)
}
