% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_uphold.R
\name{format_uphold}
\alias{format_uphold}
\title{Format Uphold file}
\usage{
format_uphold(data)
}
\arguments{
\item{data}{The dataframe}
}
\description{
Format a .csv transaction history file from Uphold for later ACB processing.
}
\examples{
\dontrun{
format_uphold(data)
}
}
\keyword{crypto}
\keyword{money}
