% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listby_coin.R
\name{listby_coin}
\alias{listby_coin}
\title{List transactions by coin}
\usage{
listby_coin(data)
}
\arguments{
\item{data}{The dataframe}
}
\description{
Provides a list of transactions, separated by coin..
}
\examples{
\dontrun{
listby_coin(formatted.ACB)
}
}
\keyword{crypto}
\keyword{money}
