% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_prices.R
\name{match_prices}
\alias{match_prices}
\title{Get Fair Market Value (FMV) of transactions}
\usage{
match_prices(data)
}
\arguments{
\item{data}{The dataframe}
}
\description{
Matches prices obtained through the \code{fetch_prices()} function with the transaction data frame.
}
\examples{
\dontrun{
match_prices(data)
}
}
\keyword{crypto}
\keyword{money}
