% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_exchanges.R
\name{merge_exchanges}
\alias{merge_exchanges}
\title{List transactions by coin}
\usage{
merge_exchanges(...)
}
\arguments{
\item{...}{To pass the other exchanges to be merged.}
}
\value{
A data frame, with rows binded and arranged, of the provided
data frames.
}
\description{
Provides a list of transactions, separated by coin..
}
\examples{
shakepay <- format_shakepay(data_shakepay)
newton <- format_newton(data_newton)
merge_exchanges(shakepay, newton)
}
