% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_exchanges.R
\name{merge_exchanges}
\alias{merge_exchanges}
\title{Merge formatted exchange transactions}
\usage{
merge_exchanges(...)
}
\arguments{
\item{...}{To pass the other exchanges to be merged.}
}
\value{
A data frame, with rows binded and arranged, of the provided
data frames.
}
\description{
Merge formatted exchange transactions into one data frame.
}
\examples{
shakepay <- format_shakepay(data_shakepay)
newton <- format_newton(data_newton)
merge_exchanges(shakepay, newton)
}
