% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_list_prices.R
\docType{data}
\name{prepare_list_prices}
\alias{prepare_list_prices}
\alias{add_popular_slugs}
\alias{popular_slugs}
\title{Prepare the list of coins for prices}
\format{
An object of class \code{data.frame} with 31 rows and 2 columns.
}
\usage{
prepare_list_prices(
  slug,
  start.date,
  end.date = lubridate::now("UTC"),
  force = FALSE
)

add_popular_slugs(data, slug_dictionary = popular_slugs)

popular_slugs
}
\arguments{
\item{slug}{Which coins to include in the list. However, because of too
many duplicated symbols, you must use the unique "slug" name of the coin.
You can obtain the correct slug using
\code{\link[crypto2:crypto_list]{crypto2::crypto_list()}} and then filtering for your
symbol.}

\item{start.date}{What date to start reporting prices for.}

\item{end.date}{What date to end reporting prices for.
the \code{list.prices} object symbol for a given coin is shared by
multiple coins (see details).}

\item{force}{Whether to force recreating \code{list.prices} even though
it already exists (e.g., if you added new coins or new dates).}

\item{data}{The data to which to convert the coin symbol / ticker to the
"slug" (full name of the coin).}

\item{slug_dictionary}{A table of equivalency between the coin symbol /
ticker and its slug. By default uses the provided \code{popular_slugs} object
with the most popular coins. This is to avoid all the duplicated coin
tickers that are returned with \code{\link[crypto2:crypto_list]{crypto2::crypto_list()}}.}
}
\value{
A data frame, with the following columns: timestamp, id, slug,
name, symbol, ref_cur, open, high, low, close, volume, market_cap,
time_open, time_close, time_high, time_low, spot.rate2, currency, date2.
}
\description{
Prepare the list of coins for prices.
}
\details{
The \link[crypto2:crypto_history]{crypto2::crypto_history} API is at times a bit capricious. You might
need to try a few times before it processes correctly and without
errors.

Sometimes, \code{list.prices} (through coinmarketcap) will contain symbols for
a given coin (e.g., ETH) that is actually shared by multiple coins, thus,
the necessity of the \code{remove.coins.slug} argument. In these cases, we can
look at the slug column of \code{list.prices} to identify the correct coins.
ETH for example possesses two slugs: "ethereum" (the main one) and
"the-infinite-garden" (probably not the one you want). Per default, a
number of duplicate slugs are excluded, the list of which is accessible in
the \code{slugs_to_remove} object. However, you can provide your own list,
should you wish to.
}
\examples{
my.coins <- c("bitcoin", "ethereum")
my.list.prices <- prepare_list_prices(slug = my.coins, start.date = "2023-01-01")
head(my.list.prices)
}
\keyword{datasets}
