% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_report.R
\name{print_report}
\alias{print_report}
\title{Print full crypto tax report}
\usage{
print_report(tax.year, name, table.revenues, report.summary, sup.losses)
}
\arguments{
\item{tax.year}{The tax year desired.}

\item{name}{Name of the individual for the report.}

\item{table.revenues}{The table of revenues.}

\item{report.summary}{The report summary.}

\item{sup.losses}{The calculated superficial losses.}
}
\description{
Will output a full crypto tax report in HTML format, which can then be printed or saved as PDF.
}
\examples{
\dontrun{
print_report()
}

}
\keyword{crypto}
\keyword{money}
