% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_overview.R
\name{report_overview}
\alias{report_overview}
\title{Summary of ACB}
\usage{
report_overview(
  formatted.ACB,
  today.data = TRUE,
  tax.year = "all",
  local.timezone = Sys.timezone()
)
}
\arguments{
\item{formatted.ACB}{The formatted ACB data.}

\item{today.data}{whether to fetch today's data.}

\item{tax.year}{Which tax year(s) to include.}

\item{local.timezone}{Which time zone to use for the date of the report.}
}
\description{
Provides a summary of Adjusted Cost Base (ACB) per share, as well as realized and unrealized gains/losses (plus net value), per coin.
}
\examples{
\dontrun{
report_overview(formatted.ACB)
}
}
\keyword{crypto}
\keyword{money}
