% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_summary.R
\name{report_summary}
\alias{report_summary}
\title{Summary of gains and losses}
\usage{
report_summary(
  formatted.ACB,
  today.data = TRUE,
  tax.year = "all",
  local.timezone = Sys.timezone()
)
}
\arguments{
\item{formatted.ACB}{The formatted ACB data.}

\item{today.data}{whether to fetch today's data.}

\item{tax.year}{Which tax year(s) to include.}

\item{local.timezone}{Which time zone to use for the date of the report.}
}
\description{
Provides a summary of realized capital gains and losses (and total).
}
\examples{
\dontrun{
report_summary(formatted.ACB)
}
}
\keyword{crypto}
\keyword{money}
