% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_box.R
\name{tax_box}
\alias{tax_box}
\title{Get a simple table of relevant tax information}
\usage{
tax_box(report.summary, sup.losses, table.revenues, proceeds)
}
\arguments{
\item{report.summary}{report.summary}

\item{sup.losses}{sup.losses}

\item{table.revenues}{table.revenues}

\item{proceeds}{proceeds}
}
\description{
Output a simple table with all the relevant tax information and tax form line numbers.
}
\examples{
\dontrun{
ACB(data)
}

}
