% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USD2CAD.R
\name{USD2CAD}
\alias{USD2CAD}
\title{Convert USD to CAD}
\usage{
USD2CAD(data, conversion = "USD", currency = "CAD")
}
\arguments{
\item{data}{The data}

\item{conversion}{What to convert to}

\item{currency}{What to convert from}
}
\value{
A data frame, with the following columns: date, CAD.rate.
}
\description{
This function allows you to convert USD to CAD.
}
\examples{
formatted.dates <- format_shakepay(data_shakepay)[1]
USD2CAD(formatted.dates)
}
