% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USD2CAD.R
\name{USD2CAD}
\alias{USD2CAD}
\title{Convert USD to CAD}
\usage{
USD2CAD(data, conversion = "USD", currency = "CAD")
}
\arguments{
\item{data}{The data}

\item{conversion}{What to convert to}

\item{currency}{What to convert from}
}
\description{
This function allows you to convert USD to CAD.
}
\examples{
# Not work anymore
# USD2CAD(date = "2021-10-08", conversion = "USD", currency = "CAD")
}
\keyword{money}
